Feat/rework

+++ cmd/client/src/stores/role.ts
@@ -2,19 +2,19 @@ import { defineStore } from 'pinia'
 import { config, logger } from '@/config'
 import { APIClient } from '@api/api.client'
 import { GrpcWebFetchTransport } from '@protobuf-ts/grpcweb-transport'
-import { CreateRoleReq, CreateRoleUserReq, DeleteRoleReq, DeleteRoleUserReq, ListRoleReq, RolePermission, RoleUsers, UpdateRoleReq } from '@internal/user/dto/role'
+import { CreateRoleReq, CreateRoleUserReq, DeleteRoleReq, DeleteRoleUserReq, ListRoleReq, ListRoleResp, RolePermission, RoleUserResp, UpdateRoleReq } from '@internal/user/dto/role'
 import { useAuthStore } from './auth'
 import { computed, ref } from 'vue'
-import { ulid } from '@/utils/ulid'
+import { ulid, zero } from '@/utils/ulid'
 import type { Permission } from '@internal/user/role'
-import type { U } from '@internal/user/user'
 
 export const useRoleStore = defineStore('role', () => {
   const roles = ref<Map<string, RolePermission>>(new Map())
   const total = ref<bigint>(BigInt(0))
 
-  // users by role
-  const users = ref<Map<string, U[]>>(new Map())
+  // roles by user
+  // user_id -> role_id -> true
+  const rolesByUser = ref<Map<string, Map<string, boolean>>>(new Map())
 
   const selected = ref<RolePermission[]>([])
 
@@ -55,34 +55,55 @@ export const useRoleStore = defineStore('role', () => {
   // Return roles ids and entity ids
   const list = async function (req: ListRoleReq): Promise<string[]> {
     try {
-      const resp = await api.listRole(req, { meta: { token: token.value } })
+      const resp: { response: ListRoleResp } = await api.listRole(req, { meta: { token: token.value } })
 
-      resp.response.roles?.forEach((role: RoleUsers) => {
-        roles.value?.set(ulid(role.role?.role?.iD), role?.role!)
-        users.value?.set(ulid(role.role?.role?.iD), role.users)
+      resp.response.roles?.forEach((role: RolePermission) => {
+        roles.value?.set(ulid(role.role?.iD), role)
       })
+
+      const roleIDs: string[] = resp.response.roles.map((role: RolePermission) => ulid(role.role?.iD))
+
+      if (req.userID.length > 0) {
+        rolesByUser.value.set(ulid(req.userID), roleIDs.reduce((acc: Map<string, boolean>, roleID: string) => {
+          acc.set(roleID, true);
+
+          return acc;
+        }, new Map<string, boolean>()))
+      }
 
       if (req?.iDs.length === 0) {
         total.value = resp.response.total
       }
 
-      return resp.response.roles.map((role: RoleUsers) => ulid(role.role?.role?.iD))
+      return roleIDs
     } catch (err: any) {
       logger.error(err)
       throw err
     }
   }
 
-  // Add user refresh cache in selected role only
+  // Add user to role
   const addUser = async function (roleID: Uint8Array, userID: Uint8Array) {
     try {
+      // zero case exception, dry update local only
+      if (ulid(userID) === ulid(zero)) {
+        const roles = rolesByUser.value.get(ulid(userID)) ?? new Map()
+        roles?.set(ulid(zero), true)
+        rolesByUser.value.set(ulid(roleID), roles)
+
+        return
+      }
+
       const req = CreateRoleUserReq.create({
         userID: userID,
         roleID: roleID
       })
 
-      const resp = await api.createRoleUser(req, { meta: { token: token.value } })
-      users.value.set(ulid(resp.response.role?.iD), resp.response.users)
+      const resp: { response: RoleUserResp } = await api.createRoleUser(req, { meta: { token: token.value } })
+
+      const roles = rolesByUser.value.get(ulid(resp.response.user?.iD)) ?? new Map()
+      roles?.set(ulid(resp.response.role?.role?.iD), true)
+      rolesByUser.value.set(ulid(resp.response.user?.iD), roles)
     } catch (err: any) {
       logger.error(err)
       throw err
@@ -92,18 +113,25 @@ export const useRoleStore = defineStore('role', () => {
 
   const deleteUser = async function (roleID: Uint8Array, userID: Uint8Array) {
     try {
+      // zero case exception, dry update local only
+      if (ulid(userID) === ulid(zero)) {
+        const roles = rolesByUser.value.get(ulid(userID)) ?? new Map()
+        roles?.delete(ulid(zero))
+        rolesByUser.value.set(ulid(userID), roles)
+
+        return
+      }
+
       const req = DeleteRoleUserReq.create({
         userID: userID,
         roleID: roleID
       })
 
-      const resp = await api.deleteRoleUser(req, { meta: { token: token.value } })
-      users.value.set(ulid(resp.response.role?.iD), resp.response.users)
+      const resp: { response: RoleUserResp } = await api.deleteRoleUser(req, { meta: { token: token.value } })
 
-      // if no users left, remove user from cache
-      if (users.value.get(ulid(resp.response.role?.iD))?.length === 0) {
-        users.value.delete(ulid(resp.response.role?.iD))
-      }
+      const roles = rolesByUser.value.get(ulid(resp.response.user?.iD)) ?? new Map()
+      roles?.delete(ulid(resp.response.role?.role?.iD))
+      rolesByUser.value.set(ulid(resp.response.user?.iD), roles)
     } catch (err: any) {
       logger.error(err)
       throw err
@@ -113,7 +141,7 @@ export const useRoleStore = defineStore('role', () => {
 
   const delete_ = async (req: DeleteRoleReq) => {
     try {
-      const resp = await api.deleteRole(req, { meta: { token: token.value } })
+      const resp: { response: RolePermission } = await api.deleteRole(req, { meta: { token: token.value } })
 
       roles.value?.delete(ulid(resp.response.role?.iD))
     } catch (err: any) {
@@ -124,14 +152,14 @@ export const useRoleStore = defineStore('role', () => {
 
   return {
     roles,
-    users,
+    rolesByUser,
     total,
     selected,
     create,
     update,
     list,
+    delete_,
     addUser,
     deleteUser,
-    delete_,
   }
 })
